requirements.txt
streamlit>=1.31.0
google-api-python-client>=2.118.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
python-dotenv>=1.0.0
pytz>=2024.1
openai>=1.12.0
streamlit-calendar>=1.1.0 